+++ examples/with-plugin/content/my-first-blog/meta.js
@@ -1,0 +1,6 @@
+module.exports = {
+  title: "My First Blog",
+  description: "This is my first blog",
+  $createdAt: "May 11 2020",
+  foo: 'Coming from JavaScript meta file'
+}