+++ examples/with-plugin/plugins/index.js
@@ -1,9 +1,16 @@
-function beforeBuild() {
-  console.log('I run before build');
+function beforeBuild(programInfo, { createContent }) {
+  const sourceNode = {
+    slug: 'hello-custom-blog',
+    markdownContent: '# Hello',
+    createdAt: new Date('13 May 2019'),
+    modifiedAt: new Date('13 May 2019')
+  }
+
+  createContent(sourceNode)
 }
 
 function afterBuild() {
-  console.log('I run after build');
+  console.log('After build working!');
 }
 
-module.exports = {beforeBuild, afterBuild}
+module.exports = { beforeBuild, afterBuild };