+++ tests/main.spec.js
@@ -5,9 +5,8 @@ describe('main()', () => {
   it('should export neccessary functions', () => {
     expect(Object.keys(main)).to.have.members([
       'build',
-      'serve',
-      'generateContentFile',
-      'generateHTMLFile'
+      'generateSite',
+      'createHTMLFile'
     ]);
   });
 });